requirements.txt
Flask==0.11.1
requests==2.12.1
simplejson==3.10.0
beautifulsoup4==4.5.1
PyYAML==3.12
